Remove the stale version of ca-certificates
Both the meta-flatpak and oe-core provide the recipes of
ca-certificates. The one in meta-flatpak is stale, but it is
still selected automatically due to the higher layer priority of
meta-flatpak. So drop this stale version from meta-flatpak. This will
also fix a warning like below due to the missing of some package
dependency.
  WARNING: seco-image-minimal-ostree-runtime-0.1-r0 do_rootfs: [log_check] seco-image-minimal-ostree-runtime: found 1 warning message in the logfile:
  [log_check] warning: %post(ca-certificates-20160104-r0.0.noarch) scriptlet failed, exit status 127

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>